requirements.txt
google-adk==0.5.0
litellm==1.70.0
python-dotenv==1.1.0
